main/requirements.txt
asgiref==3.6.0
attrs==22.2.0
black==23.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
coverage==7.2.1
cryptography==39.0.2
defusedxml==0.7.1
Django==4.1.7
django-allauth==0.53.1
django-ckeditor==6.5.1
django-froala-editor==4.0.17
django-htmx==1.14.0
django-js-asset==2.0.0
django-taggit==3.1.0
exceptiongroup==1.1.0
factory-boy==3.2.1
Faker==17.6.0
flake8==6.0.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
mccabe==0.7.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.0
pathspec==0.11.0
Pillow==9.4.0
platformdirs==3.1.0
pluggy==1.0.0
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.6.0
pytest==7.2.2
pytest-cov==4.0.0
pytest-django==4.5.2
pytest-factoryboy==2.5.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.5.0
tzdata==2022.7
urllib3==1.26.15
